modified

+++ src/main/java/com/scbb/bank/savings/service/SavingsService.java
@@ -1,5 +1,6 @@
 package com.scbb.bank.savings.service;
 
+import com.scbb.bank.exception.ResourceCannotDeleteException;
 import com.scbb.bank.exception.ResourceNotFoundException;
 import com.scbb.bank.interfaces.AbstractService;
 import com.scbb.bank.ledger.service.AccountService;
@@ -10,6 +11,8 @@ import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -40,14 +43,20 @@ public class SavingsService implements AbstractService<Savings, Integer> {
 
 	@Transactional
 	public Savings persist(Savings savings) {
-		if (savings.getId() == null)
+		if (savings.getId() == null) {
 			accountService.persist(savings.getAccount());
+			savings.setOpenedDate(LocalDate.now());
+		}
 		return savingsRepository.save(savings);
 	}
 
 	@Transactional
 	public void delete(Integer id) {
-		savingsRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Resource having id " + id + " cannot find"));
+		Savings savings = savingsRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Resource having id " + id + " cannot find"));
+		if (savings.getAccount().getBalance().compareTo(new BigDecimal("0")) != 0) {
+			throw new ResourceCannotDeleteException("Savings having id " + id + " cannot be deleted");
+		}
 		savingsRepository.deleteById(id);
 	}
